requirements.txt
Django==2.2
django-filter==2.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
Pillow==6.2.1
pytz==2019.2
sqlparse==0.3.0
